requirements.txt
setuptools
gitpython